backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
supabase==2.10.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.17
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
networkx==3.4.2
openai==1.58.1
python-dateutil==2.9.0.post0
email-validator==2.2.0